refactor: move all multiSig IPEX activity to ipexCommunicationService
